align openvino namespace
