+++ tools/cross_check_tool/openvino/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (C) 2018-2021 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+__path__ = __import__('pkgutil').extend_path(__path__, __name__)
