+++ tools/cross_check_tool/openvino/tools/cross_check_tool/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (C) 2018-2021 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+
