+++ tools/cross_check_tool/openvino/tools/cross_check_tool/utils.py
